Get exam types from Firebase properly

+++ app/src/main/java/io/diaryofrifat/code/examroutine/data/local/ExamType.kt
@@ -1,5 +1,5 @@
 package io.diaryofrifat.code.examroutine.data.local
 
-data class ExamType(var id: Long,
+data class ExamType(var id: Int,
                     val examTypeKey: String,
                     val examTypeTitle: String)
